Add files via upload
This is all the informmation for the airlines. I had to filter out all the added information so that I could correctly assign primariy keys to the tables
I also did the same with the airports so that I could later add them into the DB and query without any mismatching keys.